Moved index, column tests into test_models #52
